Update README.md
This Gym Management System project is a console application developed in C++ programming language, employing Object-Oriented techniques. Its main objective is to provide a system that efficiently manages the data of individuals entering a gym center and maintains their health records. The project involves several actors or users, including the Admin, Member, and Trainer, each with their respective roles and responsibilities. The entities involved in the system are Gym, Member, Trainer, and Admin, which interact to facilitate seamless management of gym operations, member information, trainer assignments, and administrative tasks. Overall, this project aims to enhance the organization and maintenance of a gym center, streamlining the processes and ensuring effective health care management for its members.